pyproject.toml
[project]
name = "CalyxFlow"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "llama-index-agent-openai>=0.4.12",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-llms-openai-like>=0.4.0",
    "llama-index-llms-vllm>=0.5.1",
    "vllm==0.12.0",
    "torch",
]

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.uv]
environments = [ "sys_platform == 'linux' and platform_machine == 'aarch64'"]

[tool.uv.sources]
torch = { index = "pytorch-cu129" }
